hikptool/RDMA: Fix the missing bank ID in roce_trp GEN_AC
Upstream: Yes

Currently hikptool only accepts bank ID 0 and 1 for roce_trp GEN_AC.
The firmware then multiplies the bank ID received from hikptool by 2
and converts it to 0 and 2. But actually there are 4 banks in GEN_AC
and each of them is independent.

To fix the problem, both hikptool and firmware should be modified.
This patch modifies the bank ID limit form 2 to 4 in hikptool.

Signed-off-by: Junxian Huang <huangjunxian6@hisilicon.com>