hikptool/roce: Fix the infinite loop due to compatibility issue
When a new-version hikptool sends a subcmd of caep ext registers to
an old-version FW which doesn't support the query of ext registers
yet, due to the lack of subcmd check in old-version FW, there will
be no error and register data will still be responsed to hikptool.

In the new-version hikptool, a do-while loop is used to implement a
multi-round query, and the loop variable is updated based on the
cur_block_num responsed from FW. Since there is no error reported,
the new-version hikptool will parse the old-version data based on the
new-version data structure. This may cause the parsed cur_block_num
remains 0, and the loop becomes infinite.

To fix this problem, add a check of the responsed cur_block_num.
If it's 0, return with an error code.

Fixes: 8b3b68347165 ("hikptool/roce: Add a common frame for hikptool roce register query")
Signed-off-by: Junxian Huang <huangjunxian6@hisilicon.com>