hikptool/roce: Fix redundant update of total_block_num
total_block_num needs to be updated only in the first loop of query,
and should stay constant from then on. Otherwise there may be
out-of-bound accessing problems if subsequent update value from FW
is larger than the first one, with which the memory of the register
array is allocated.

Fixes: 8b3b68347165 ("hikptool/roce: Add a common frame for hikptool roce register query")
Signed-off-by: Junxian Huang <huangjunxian6@hisilicon.com>