hikptool: Added the imp subcommand to support dumping firmware logs.
A new imp subcommand has been added, which uses the
'-l/--log' option to specify the function of dumping
firmware logs. Additionally, the '-c' and '-d' options
must be used to specify the corresponding chip and die
to be dumped.

Usage: hikptool imp -l -c xx -d xx.

Note: If an invalid or unsupported chip or die is entered,
the command will report an error.

Signed-off-by: veega2022 <zhuweijia@huawei.com>