hikptool: Added the scc subcommand to support querying DFX information.
A new scc subcommand has been added, which uses the
'-l/--log' option to specify the function of dumping
firmware logs. Additionally, the '-c' and '-d' options
must be used to specify the corresponding chip and die
to be dumped.

Usage: hikptool scc -l -c xx -d xx.

The '-v' option is also supported for querying the
firmware version number. When querying,
the '-c' and '-d' options must also be included.

Usage: hikptool scc -v -c xx -d xx.

Note: If an invalid or unsupported chip or die is entered,
the command will report an error.

Signed-off-by: veega2022 <zhuweijia@huawei.com>